pyproject.toml
[tool.poetry]
name = "tap-ethereum"
version = "0.0.1"
description = "`tap-ethereum` is a Singer tap for Ethereum, built with the Meltano SDK for Singer Taps."
authors = ["Matt Evenson"]
keywords = [
    "ELT",
    "Ethereum",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.2"
requests = "^2.25.1"
singer-sdk = "^0.4.6"
web3 = "^5.29.0"
etherscan-python = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_ethereum"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-ethereum = 'tap_ethereum.tap:TapEthereum.cli'
